refactor: move ImmutableKnotVector into knotspace.py
